Opower branch setup
This branch will be used to build the badgeville_berlin_opower gem.
The gem will live alongside the badgeville_berlin gem in our gemserver.

It is to be used internally at Opower whenever we have pending pull requests
with Badgeville but have a need to run those code changes prior to the acceptance
of the pull requests.